Documentation update -- more / better examples
Plus: Dropping util.py and moving simple_auth_handler into it.
This gives us a place for future 'util' code to live, plus
other structural benefits.
