+++ paramiko_jump/__init__.py
@@ -4,10 +4,8 @@ paramiko_jump: Paramiko + Jump Host + Multi-Factor Authentication
 
 """
 
-from paramiko_jump.client import (
-    SSHJumpClient,
-    simple_auth_handler,
-)
+from paramiko_jump.client import SSHJumpClient
+from paramiko_jump.util import simple_auth_handler
 
 __all__ = (
     'SSHJumpClient',
