+++ paramiko_jump/client.py
@@ -6,13 +6,10 @@ multi-factor-authentication.
 
 """
 
-from getpass import getpass
 from typing import (
     AnyStr,
     Callable,
-    List,
     Optional,
-    Sequence,
     Tuple,
     Union,
 )
@@ -22,38 +19,6 @@ from paramiko.client import SSH_PORT, SSHClient
 
 Host = Union[AnyStr, Tuple[AnyStr, int]]
 Prompt = Tuple[AnyStr, bool]
-
-
-def simple_auth_handler(
-        title: AnyStr,
-        instructions: AnyStr,
-        prompt_list: Sequence[Prompt],
-) -> List[AnyStr]:
-    """
-    Authentication callback, for keyboard-interactive
-    authentication.
-
-    :param title:
-        Displayed to the end user before anything else.
-    :param instructions:
-        Displayed to the end user. Typically contains text explaining
-        the authentication scheme and / or legal disclaimers.
-    :param prompt_list:
-        A Sequence of (AnyStr, bool). Each string element is
-        displayed as an end-user input prompt. The corresponding
-        boolean element indicates whether the user input should
-        be 'echoed' back to the terminal during the interaction.
-    """
-    answers = []
-    if title:
-        print(title)
-    if instructions:
-        print(instructions)
-
-    for prompt, show_input in prompt_list:
-        input_ = input if show_input else getpass
-        answers.append(input_(prompt))
-    return answers
 
 
 class SSHJumpClient(SSHClient):
